requirements.txt
coverage>=3.7.1
flake8>=2.1.0
Django>=2.0.5
django-filter>=1.1.0
django-nose>=1.4.5
djangorestframework>=3.8.2
plotly